This is the first form of a small program that allows the client to connect remotely and run a script on the server computer. The objective of this proyect is to create a tool for my friends to run ciertan scripts I made on my server without them having to know a lot of technical stuff. This first commit is more of a proof of concept right now.
Right now it can:
Run a server client that opens a socket
Run a client that conects to said server
send a message to the server that triggers it to run a script.

In this commit also comes an example of how to run a bat file in java.
The name of that class is runscript.java (how creative).